requirements/ci.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements/ci.txt requirements/base.txt requirements/ci.in requirements/testing.in
#
alabaster==0.7.12
    # via
    #   -r requirements/base.txt
    #   sphinx
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
astroid==2.11.7
    # via pylint
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
babel==2.6.0
    # via
    #   -r requirements/base.txt
    #   sphinx
black==22.6.0
    # via -r requirements/testing.in
bleach==5.0.1
    # via
    #   -r requirements/base.txt
    #   django-markup
certifi==2018.4.16
    # via
    #   -r requirements/base.txt
    #   requests
chardet==3.0.4
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via black
codecov==2.1.12
    # via -r requirements/ci.in
coverage==6.4.4
    # via
    #   -r requirements/ci.in
    #   codecov
dill==0.3.5.1
    # via pylint
django==3.2.13
    # via
    #   -r requirements/base.txt
    #   django-appconf
    #   django-axes
    #   django-capture-on-commit-callbacks
    #   django-choices
    #   django-filter
    #   django-markup
    #   django-privates
    #   django-rest-framework-condition
    #   django-sendfile2
    #   djangorestframework
    #   drf-extra-fields
    #   drf-nested-routers
    #   drf-spectacular
    #   vng-api-common
django-appconf==1.0.5
    # via -r requirements/base.txt
django-axes==5.33.0
    # via -r requirements/base.txt
django-capture-on-commit-callbacks==1.11.0
    # via -r requirements/testing.in
django-choices==1.7.2
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-cors-middleware==1.3.1
    # via -r requirements/base.txt
django-filter==21.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-ipware==4.0.2
    # via
    #   -r requirements/base.txt
    #   django-axes
django-markup[all_filter_dependencies]==1.5
    # via -r requirements/base.txt
django-privates==1.0.1.post0
    # via -r requirements/base.txt
django-rest-framework-condition==0.1.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-sendfile2==0.4.2
    # via
    #   -r requirements/base.txt
    #   django-privates
django-solo==1.1.3
    # via
    #   -r requirements/base.txt
    #   vng-api-common
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   drf-extra-fields
    #   drf-nested-routers
    #   drf-spectacular
    #   vng-api-common
djangorestframework-camel-case==1.3.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
docutils==0.14
    # via
    #   -r requirements/base.txt
    #   django-markup
    #   python-creole
    #   sphinx
drf-extra-fields==3.4.0
    # via -r requirements/base.txt
drf-nested-routers==0.93.4
    # via
    #   -r requirements/base.txt
    #   vng-api-common
drf-spectacular==0.24.2
    # via
    #   -r requirements/base.txt
    #   vng-api-common
drf-writable-nested==0.6.3
    # via -r requirements/base.txt
factory-boy==3.2.1
    # via -r requirements/testing.in
faker==14.1.0
    # via factory-boy
freezegun==1.2.2
    # via -r requirements/testing.in
gemma-zds-client==1.0.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
html5lib==1.1
    # via
    #   -r requirements/base.txt
    #   textile
humanize==0.5.1
    # via -r requirements/base.txt
idna==2.7
    # via
    #   -r requirements/base.txt
    #   requests
imagesize==1.1.0
    # via
    #   -r requirements/base.txt
    #   sphinx
inflection==0.3.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
iso-639==0.4.5
    # via
    #   -r requirements/base.txt
    #   vng-api-common
isodate==0.6.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
isort==5.10.1
    # via
    #   -r requirements/testing.in
    #   pylint
jinja2==2.10.1
    # via
    #   -r requirements/base.txt
    #   sphinx
jsonschema==4.16.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
lazy-object-proxy==1.7.1
    # via astroid
markdown==3.1
    # via
    #   -r requirements/base.txt
    #   django-markup
markupsafe==1.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
mccabe==0.7.0
    # via pylint
mypy-extensions==0.4.3
    # via black
oyaml==0.7
    # via
    #   -r requirements/base.txt
    #   vng-api-common
packaging==19.0
    # via
    #   -r requirements/base.txt
    #   sphinx
pathspec==0.9.0
    # via black
pep8==1.7.1
    # via -r requirements/testing.in
pillow==6.2.1
    # via -r requirements/base.txt
platformdirs==2.5.2
    # via
    #   black
    #   pylint
psycopg2-binary==2.9.3
    # via -r requirements/base.txt
pygments==2.3.1
    # via
    #   -r requirements/base.txt
    #   django-markup
    #   sphinx
pyjwt==1.6.4
    # via
    #   -r requirements/base.txt
    #   gemma-zds-client
    #   vng-api-common
pylint==2.14.5
    # via -r requirements/testing.in
pyparsing==2.3.1
    # via
    #   -r requirements/base.txt
    #   packaging
pyrsistent==0.18.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
python-creole==1.4.10
    # via
    #   -r requirements/base.txt
    #   django-markup
python-dateutil==2.7.3
    # via
    #   -r requirements/base.txt
    #   faker
    #   freezegun
python-dotenv==0.8.2
    # via -r requirements/base.txt
pytz==2018.4
    # via
    #   -r requirements/base.txt
    #   babel
    #   django
pyyaml==5.1.2
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
raven==6.9.0
    # via -r requirements/base.txt
regex==2022.7.25
    # via
    #   -r requirements/base.txt
    #   textile
requests==2.21.0
    # via
    #   -r requirements/base.txt
    #   codecov
    #   gemma-zds-client
    #   requests-mock
    #   sphinx
    #   vng-api-common
requests-mock==1.9.3
    # via -r requirements/testing.in
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   bleach
    #   django-choices
    #   django-sendfile2
    #   html5lib
    #   isodate
    #   packaging
    #   python-dateutil
    #   requests-mock
    #   sphinx
smartypants==2.0.1
    # via
    #   -r requirements/base.txt
    #   django-markup
snowballstemmer==1.2.1
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinx==1.8.4
    # via
    #   -r requirements/base.txt
    #   sphinx-rtd-theme
sphinx-rtd-theme==0.4.2
    # via -r requirements/base.txt
sphinxcontrib-websupport==1.1.0
    # via
    #   -r requirements/base.txt
    #   sphinx
sqlparse==0.3.0
    # via
    #   -r requirements/base.txt
    #   django
tblib==1.7.0
    # via -r requirements/testing.in
textile==4.0.2
    # via
    #   -r requirements/base.txt
    #   django-markup
tomli==2.0.1
    # via
    #   black
    #   pylint
tomlkit==0.11.4
    # via pylint
typing-extensions==4.3.0
    # via
    #   astroid
    #   black
    #   pylint
uritemplate==3.0.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==1.24.3
    # via
    #   -r requirements/base.txt
    #   requests
vng_api_common @ git+https://github.com//MatthijsBekendam/vng-api-common.git@drf_compatibilty_newer_django_fix
    # via -r requirements/base.txt
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
    #   html5lib
wrapt==1.14.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
